+++ src/models/filament.js
@@ -423,6 +423,22 @@ module.exports = function (app) {
         });
     };
 
+    filamentSchema.statics.getCountPerShopsMapping = async function () {
+        let results = await this.aggregate([
+            { $group: {
+                    _id: '$shop',
+                    count: { $sum: 1 }
+                }
+            },
+            { $sort: { 'count': -1 } }
+        ]).exec();
+
+        return results.reduce((mapping, doc) => {
+            mapping[doc._id] = doc.count;
+            return mapping;
+        }, {});
+    };
+
     filamentSchema.statics.getCountPerBrandsMapping = async function () {
         let results = await this.aggregate([
             { $group: {