+++ src/models/shop.js
@@ -26,5 +26,38 @@ module.exports = function (app) {
         return this.findOne({}, {}, {skip: rand}, callback);
     };
 
+    shopSchema.statics.list = async function (options) {
+        const Filament = app.models.filament;
+
+        options = Object.assign({
+            countFilaments: false,
+            locale: null
+        }, options);
+
+        let results = await this.find().sort('name').exec();
+        if (options.countFilaments) {
+            let filamentsCounts = await Filament.getCountPerShopsMapping();
+            for (let shop of results) {
+                if (filamentsCounts[shop.id]) {
+                    shop.filamentsCount = filamentsCounts[shop.id];
+                } else {
+                    shop.filamentsCount = 0;
+                }
+            }
+        }
+
+        return results;
+    };
+
+    shopSchema.pre('remove', async function(next) {
+        const Filament = app.models.filament;
+
+        if (await Filament.find({shop: this.id}).countDocuments().exec() > 0) {
+            return next(new Error("A shop used by filament(s) can not be deleted."));
+        } else {
+            return next();
+        }
+    });
+
     return mongoose.model('Shop', shopSchema);
 };